category service test

+++ AuthenticatorPro.Core/src/Service/Impl/CategoryService.cs
@@ -1,7 +1,6 @@
 // Copyright (C) 2022 jmh
 // SPDX-License-Identifier: GPL-3.0-only
 
-using AuthenticatorPro.Core.Comparer;
 using AuthenticatorPro.Core.Entity;
 using AuthenticatorPro.Core.Persistence;
 using AuthenticatorPro.Core.Persistence.Exception;
@@ -16,16 +15,29 @@ namespace AuthenticatorPro.Core.Service.Impl
     {
         private readonly ICategoryRepository _categoryRepository;
         private readonly IAuthenticatorCategoryRepository _authenticatorCategoryRepository;
+        private readonly IEqualityComparer<Category> _equalityComparer;
 
         public CategoryService(ICategoryRepository categoryRepository,
-            IAuthenticatorCategoryRepository authenticatorCategoryRepository)
+            IAuthenticatorCategoryRepository authenticatorCategoryRepository,
+            IEqualityComparer<Category> equalityComparer)
         {
             _categoryRepository = categoryRepository;
             _authenticatorCategoryRepository = authenticatorCategoryRepository;
+            _equalityComparer = equalityComparer;
         }
 
         public async Task TransferAsync(Category initial, Category next)
         {
+            if (initial == null)
+            {
+                throw new ArgumentException("Initial category cannot be null");
+            }
+
+            if (next == null)
+            {
+                throw new ArgumentException("Next category cannot be null");
+            }
+
             await _categoryRepository.CreateAsync(next);
             await _authenticatorCategoryRepository.TransferAsync(initial, next);
             await _categoryRepository.DeleteAsync(initial);
@@ -35,7 +47,7 @@ namespace AuthenticatorPro.Core.Service.Impl
         {
             if (categories == null)
             {
-                return 0;
+                throw new ArgumentException("Categories cannot be null");
             }
 
             var added = 0;
@@ -59,6 +71,11 @@ namespace AuthenticatorPro.Core.Service.Impl
 
         public async Task<ValueTuple<int, int>> AddOrUpdateManyAsync(IEnumerable<Category> categories)
         {
+            if (categories == null)
+            {
+                throw new ArgumentException("Categories cannot be null");
+            }
+
             var list = categories.ToList();
             var added = await AddManyAsync(list);
             var updated = await UpdateManyAsync(list);
@@ -70,17 +87,16 @@ namespace AuthenticatorPro.Core.Service.Impl
         {
             if (categories == null)
             {
-                return 0;
+                throw new ArgumentException("Categories cannot be null");
             }
 
             var updated = 0;
-            var comparer = new CategoryComparer();
 
             foreach (var category in categories)
             {
                 var original = await _categoryRepository.GetAsync(category.Id);
 
-                if (original == null || comparer.Equals(original, category))
+                if (original == null || _equalityComparer.Equals(original, category))
                 {
                     continue;
                 }
@@ -94,6 +110,11 @@ namespace AuthenticatorPro.Core.Service.Impl
 
         public async Task DeleteWithCategoryBindingsASync(Category category)
         {
+            if (category == null)
+            {
+                throw new ArgumentException("Category cannot be null");
+            }
+
             await _categoryRepository.DeleteAsync(category);
             await _authenticatorCategoryRepository.DeleteAllForCategoryAsync(category);
         }
